third_party/WebKit/LayoutTests/http/tests/push_messaging/permission-state-exception-in-document.html

<!DOCTYPE html>
<html>
<head>
<title>Push API: permissionState fails with NotSupportedError if userVisibleOnly is not used</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    if (window.testRunner)
        testRunner.setPermission('push-messaging', 'prompt', location.origin, location.origin);

    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            // If running manually, clear permissions before running this test.
            return serviceWorkerRegistration.pushManager.permissionState();
        })
        .then(function(permissionStatus) {
            assert_unreached('permissionStatus should have failed with a NotSupported exception');
        }, function(error) {
            assert_equals(error.name, 'NotSupportedError');
            assert_equals(error.message,
                          'Push subscriptions that don\'t enable userVisibleOnly are not supported.');
            test.done();
        }).catch(unreached_rejection(test));
}, 'Push API: permissionState fails with NotSupportedError if userVisibleOnly is not used');
</script>
</body>
</html>
